refactor: move global to provider folder
